Refactor + put redaction in NonPersistedEventEntry
