+++ src/matrix/room/timeline/entries/EventEntry.js
@@ -26,7 +26,6 @@ export class EventEntry extends BaseEventEntry {
         this._decryptionResult = null;
         this._contextEntry = null;
         this._contextForEntries = null;
-        this._markedAsRedacted = false;
     }
 
     clone() {
@@ -56,10 +55,6 @@ export class EventEntry extends BaseEventEntry {
             this._contextForEntries = [];
         }
         this._contextForEntries.push(entry);
-    }
-
-    setAsRedacted() {
-        this._markedAsRedacted = true;
     }
 
     get contextForEntries() {
@@ -160,7 +155,7 @@ export class EventEntry extends BaseEventEntry {
     }
 
     get isRedacted() {
-        return this._markedAsRedacted || super.isRedacted || isRedacted(this._eventEntry.event);
+        return super.isRedacted || isRedacted(this._eventEntry.event);
     }
 
     get redactionReason() {
