+++ src/matrix/room/timeline/persistence/RelationWriter.js
@@ -17,6 +17,7 @@ limitations under the License.
 import {EventEntry} from "../entries/EventEntry.js";
 import {REDACTION_TYPE, isRedacted} from "../../common.js";
 import {ANNOTATION_RELATION_TYPE, getRelation} from "../relations.js";
+import {redactEvent} from "../common.js";
 
 export class RelationWriter {
     constructor({roomId, ownUserId, fragmentIdComparer}) {
@@ -127,21 +128,7 @@ export class RelationWriter {
         // check if we're the target of a relation and remove all relations then as well
         txn.timelineRelations.removeAllForTarget(this._roomId, redactedEvent.event_id);
 
-        for (const key of Object.keys(redactedEvent)) {
-            if (!_REDACT_KEEP_KEY_MAP[key]) {
-                delete redactedEvent[key];
-            }
-        }
-        const {content} = redactedEvent;
-        const keepMap = _REDACT_KEEP_CONTENT_MAP[redactedEvent.type];
-        for (const key of Object.keys(content)) {
-            if (!keepMap?.[key]) {
-                delete content[key];
-            }
-        }
-        redactedEvent.unsigned = redactedEvent.unsigned || {};
-        redactedEvent.unsigned.redacted_because = redactionEvent;
-
+        redactEvent(redactionEvent, redactedEvent);
         delete redactedStorageEntry.annotations;
 
         return true;
@@ -223,35 +210,6 @@ function isObjectEmpty(obj) {
     return true;
 }
 
-// copied over from matrix-js-sdk, copyright 2016 OpenMarket Ltd
-/* _REDACT_KEEP_KEY_MAP gives the keys we keep when an event is redacted
- *
- * This is specified here:
- *  http://matrix.org/speculator/spec/HEAD/client_server/latest.html#redactions
- *
- * Also:
- *  - We keep 'unsigned' since that is created by the local server
- *  - We keep user_id for backwards-compat with v1
- */
-const _REDACT_KEEP_KEY_MAP = [
-    'event_id', 'type', 'room_id', 'user_id', 'sender', 'state_key', 'prev_state',
-    'content', 'unsigned', 'origin_server_ts',
-].reduce(function(ret, val) {
-    ret[val] = 1; return ret;
-}, {});
-
-// a map from event type to the .content keys we keep when an event is redacted
-const _REDACT_KEEP_CONTENT_MAP = {
-    'm.room.member': {'membership': 1},
-    'm.room.create': {'creator': 1},
-    'm.room.join_rules': {'join_rule': 1},
-    'm.room.power_levels': {'ban': 1, 'events': 1, 'events_default': 1,
-                            'kick': 1, 'redact': 1, 'state_default': 1,
-                            'users': 1, 'users_default': 1,
-                           },
-    'm.room.aliases': {'aliases': 1},
-};
-// end of matrix-js-sdk code
 
 import {createMockStorage} from "../../../../mocks/Storage";
 import {createEvent, withTextBody, withRedacts, withContent} from "../../../../mocks/event.js";
